+++ docs/api/changelog.rst
@@ -54,6 +54,8 @@ Version 0.6
 
   This provides more fail-safe behaviour while still providing enough feedback.
 
+* New method: :meth:`aioxmpp.stream.StanzaStream.send_and_wait_for_sent`
+
 
 Version 0.5
 ===========